Added selection_action and its test
